requirements.txt
jupyter==1.0.0
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
scikit-learn==1.1.3
seaborn==0.12.1
torch==1.13.0
torchvision==0.14.0
torchcontrib==0.0.2
wandb==0.13.5
